requirements.txt
colorama==0.4.6
dill==0.3.9
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.0
pygame-ce==2.5.2
pygame_gui==0.6.12
Pygments==2.18.0
python-i18n==0.3.9
rich==13.9.4
taichi==1.7.2
